Put queryDatabase within LimitOffsetPagingSource
